refactor(libs/types): extract DocFrontmatter from /build
